Make cache post job always run
While the logic might simply be "we don't want to keep the cache if
something failed", far more often you actually might. Because of how
this cache currently works (moving the directory on cache restore), not
moving it back and the end of the job effectively invalidates the cache.
It is far better to keep the cache, and let the user decide somehow if
the cache should be invalidated.